consolidate to a single Presneter
